Bump Version to 1.0.0
Wow, 1.0! This is the same as before, but I feel like it’s stable and
can be used in production.

The biggest change is that we are now using **Faker 2.0**. This update
breaks string pattern names, as the category is now written in lower
case.

Basically, when you were using `Lorem.sentence` or
`Name.firstName` before , you now need to use `lorem.sentence` and
`name.firstName`.